Changes due to renaming mpas-jedi repo (#33)
TYPE: enhancement

KEYWORDS: mpas-jedi

DESCRIPTION OF CHANGES: On 2 July 2019, the JCSDA/mpas repository was renamed to JCSDA/mpas-jedi in order to reduce naming ambiguity with the MPAS model repository. The changes in this commit are due to that renaming. Users should be aware that these changes will change the name of the code & build directories used for the mpas-jedi repository in the bundle from 'mpas' to 'mpas-jedi'.

LIST OF MODIFIED FILES:
M CMakeLists.txt
M build_mpas.csh

TESTS CONDUCTED: mpas-bundle was rebuilt in Singularity using these changed files and all ctests passed.