Add Github PR template (#37)
TYPE: feature

KEYWORDS: github, template

DESCRIPTION OF CHANGES: This adds a Github pull request template file to this repository, identical to the one in use on the mpas-jedi repository.

LIST OF MODIFIED FILES:
A .github/pull_request_template.md

TESTS CONDUCTED: Due to the nature of the change, this cannot really be tested until it is merged. But it is not a code change and cannot break anything.