Updated for plotting path and libraries on Cheyenne HPC. (#41)
TYPE: enhancement

KEYWORDS: mpas-jedi, python3

DESCRIPTION OF CHANGES: (1)On 2 July 2019, the JCSDA/mpas repository was renamed to JCSDA/mpas-jedi. The change for plotting directory in this commit is due to the renaming.
(2)Considering an end of life for Python 2 support in 2020, I switched to use python3 on Cheyenne.

LIST OF MODIFIED FILES:
build_mpas.csh

TESTS CONDUCTED: Tested on Cheyenne.