Update (commented) eckit project line (#45)
TYPE: bug fix/no impact

KEYWORDS: eckit, CMakeLists

DESCRIPTION OF CHANGES: The CMakeLists.txt file contains a commented line for building eckit. It is commented-out because eckit is usually provided by JCSDA modules or a container and so is not built as part of the bundle. However, in October JCSDA changed the version of eckit used, and this PR updates the commented line so that it is correct in case it is ever uncommented.

ISSUE:

Fixes [https://github.com/orgs/JCSDA/teams/jedi/discussions/32]

LIST OF MODIFIED FILES:
M CMakeLists.txt

TESTS CONDUCTED: Built bundle and executed ctests on Cheyenne using the uncommented line.