Update build script for new mpas-specific cheyenne modules (#46)
TYPE: enhancement

KEYWORDS: build script, cheyenne modules, gnu, intel

DESCRIPTION OF CHANGES:
First, the default build is changed to debug in order to catch more bugs during our testing. Also, the Debug mode for the libraries (PIO and Mpas-Model) is now set separately from the mpas-bundle.

The way the script selects the MPAS-Model git branch is improved to avoid an error that caused an incorrect branch to be selected.

An option is added to allow customized PIO builds, which is only used for debugging at this time.

The job for ctests on cheyenne is changed to run on 16 cores so that up to the 16pe tests can be selected in mpas-jedi/test/CMakeLists.txt. Also the portion of the build script that generates the test script is moved to the test_mpas section.

Before building the bundle, a simple check is performed to make sure that the PIO and Mpas-Model libraries are present.

ISSUE:
Related to, but does not close JCSDA/mpas-jedi#201.

LIST OF MODIFIED FILES:
M build_mpas.csh

TESTS CONDUCTED: Both gnu-openmpi and intel-impi build on Cheyenne. Only gnu-openmpi passes ctests so far. Also tested with the latest gnu-openmpi Singularity image. (A bug was discovered in the Singularity image, so the latest one (1/21/20) is required.)