Feature/cmake update (#64)
TYPE: enhancement

KEYWORDS: cmake bundle build of MPAS model

DESCRIPTION OF CHANGES:

This is the new mpas-bundle with the capability to build the MPAS model in the bundle, as a shared library. This is working and passing ctests (mostly) on Cheyenne, S4, containers, and general linux and mac installs with jedi-stack.

I will be adding two more PRs with the feature/cmake-update branch. One to mpas-jedi, the other from JCSDA/MPAS-Model fork to jjguerrette/MPAS-Model/develop-for-jedi. This was the only way I can set up my PRs to integrate with JCSDA zenhub epics. Let's perhaps set a meetting with @mmiesch to dissuss on how to manage the MPAS-Model fork long term. For now this will get it working.

Files changed:
M .gitignore
M CMakeLists.txt
D build_mpas.csh
R100 run_odb_build.sh scripts/run_odb_build.sh
R100 setup_odb_build.sh scripts/setup_odb_build.sh

TESTS CONDUCTED: Ctests
We are seeing a few minor variations in the ctest output. This appears repeatable on Cheyenne and on other systems. This may be related to open issue mpas-jedi # 376, which I have patched around in the MPAS-Model branch. A PR is in progress to adapt the tolerances of the failing tests.