Add environment setup scripts (#73)
TYPE: feature

KEYWORDS: environment setup

DESCRIPTION OF CHANGES: Due to the recent removal of the mpas-bundle build script in #64, it is now desirable to have environment setup files to run before building the mpas-bundle with ecbuild. This adds four such files for use on Cheyenne. There is one script each for gnu-openmpi and intel-impi using both sh and csh scripts.

ISSUE:

Addresses #68
LIST OF MODIFIED FILES:
A env-setup/gnu-openmpi-cheyenne.csh
A env-setup/gnu-openmpi-cheyenne.sh
A env-setup/intel-impi-cheyenne.csh
A env-setup/intel-impi-cheyenne.sh

TESTS CONDUCTED: Tested on cheyenne by running each of the scripts and verifying that the desired environment resulted.