Change MPAS-Model cmake project name to "MPAS" to match repo CMakeLists.txt (#94)
TYPE: bug fix

DESCRIPTION OF CHANGES:
There are several separate issues described in #90.  One of them is the issuance of the following error message during forecasts initialized from an "init" type file when `core_init_atmosphere` is built with cmake:

```
CRITICAL ERROR: NaN detected in 'w' field.
```
Changing the mpas-bundle project name to `MPAS`, as is done in this PR, _may_ contribute to fixing that issue for a 120km domain cycling experiment.

It may also be important to set `io_type="pnetcdf,cdf5"` in all `streams.atmosphere` (for DA, HofX, and forecast applications) and `streams.init_atmosphere` (for init_atmosphere) `streams` nodes.  I added the `io_type` setting out of an abundance of caution to ensure consistency between `init_atmosphere` and dependent applications, but do not know the actual impact other than ignoring size limits for large netcdf variables.

An additional precaution that should be followed is that for 2-stream input, the static file should be set to one of the init files that is generated with the same version of the MPAS-Model code that is used in all of the DA, HofX, and forecast applications.

ISSUE:
Aimed at #90 

LIST OF MODIFIED FILES:
M       .gitignore
M       CMakeLists.txt

TESTS CONDUCTED: For 120km domain,
 - used `mpas_init_atmosphere` to generate init files and initial "mpasout" (`da_state`) forecast files
 - Used `mpasjedi_eda.x` to perform an eda_3denvar analysis with "mpasout" files as background and one init file from above linked as the "static" input
 - Used `mpas_atmosphere` to perform 6-hr forecasts from five ensemble member from analyses.  Used same init file from above as the "static" input.